% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_get_analysis_z_scores.R
\name{get_analysis_z_scores}
\alias{get_analysis_z_scores}
\title{step3_get_analysis_z_scores}
\usage{
get_analysis_z_scores(
  TENET_directory,
  hypermeth_analysis,
  hypometh_analysis,
  usecaseonly,
  TF_only,
  significant_p_value,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step2' subdirectory and results created by the step2_get_diffmeth_regions function. This function will also create a new 'step3' subdirectory there containing the results of this function.}

\item{hypermeth_analysis}{Set to TRUE/FALSE depending on if you want to calculate z-scores for hypermethylated probes.}

\item{hypometh_analysis}{Set to TRUE/FALSE depending on if you want to calculate z-scores for hypomethylated probes.}

\item{usecaseonly}{Set to TRUE/FALSE depending on if you want to include the control/normal samples with the experimental/tumor samples when identifying hyper/hypomethylated groups and calculating z-scores.}

\item{TF_only}{Set to TRUE/FALSE to determine if you only want to consider genes that are accepted transcription factors in "The Human Transcription Factors" by Lambert et al. 2018 when calculating z-scores.}

\item{significant_p_value}{Set a p-value to identify significant Z-scores for gene expression values selected between hyper/hypomethylated tumor/experimental samples and those that are not. Defaults to 0.05.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns tab-delimited 'sig_link_zscores.txt' files for hypo/hyper Gplus/Gminus probe-gene links, as well as individual "zscores.txt" files named after each gene in the hypo/hyper analysis with z-scores for that gene to all probes from that analysis type.
}
\description{
This is the step3 function of the TENETR package.
This function takes the identified hyper and hypomethylated probes
from step2_get_diffmeth_regions function and calculates a zscore comparing
the mean expression of each gene in the experimental/tumor samples that are
hyper or hypomethylated for each probe, according to hypermeth and hypometh cutoffs
set during the previous function, to those that are not, and permutates this
across all hyper or hypomethylated probes depending on user input, calculating
a z-score for each probe and gene combination, aka link.
}
