% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diffmeth_regions.R
\name{get_diffmeth_regions}
\alias{get_diffmeth_regions}
\title{get_diffmeth_regions}
\usage{
get_diffmeth_regions(
  TENET_directory,
  TSS_dist,
  methcutoff,
  hypomethcutoff,
  unmethcutoff,
  hypermethcutoff,
  minExp
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains the step1 directory created by make_external_datasets function as well as the .rda file containing methylation and expression data.}

\item{TSS_dist}{Set a number to be the buffer in base pairs from GENCODE v22-annotated transcription start sites for DNA methylation probes to not be considered enhancer probes.}

\item{methcutoff}{Set a number from 0 to 1 to be the beta-value cutoff for methylated probes.}

\item{hypomethcutoff}{Set a number from 0 to 1 to be the beta-value cutoff for hypomethylated probes. Should be set lower than the methcutoff.}

\item{unmethcutoff}{Set a number from 0 to 1 to be the beta-value cutoff for unmethylated probes.}

\item{hypermethcutoff}{Set a number from 0 to 1 to be the beta-value cutoff for hypermethylated probes. Should be set lower than the unmethcutoff.}

\item{minExp}{Sets the minimum number of experimental/tumor samples to be considered for the hypo/hypermethylated groups.}
}
\value{
Returns several objects including a .rda file with matrices of methylation data for the four quadrants in control/normal and experimental/tumor samples. Also outputs a simple .txt file containing metrics on the number of probes found in different categories.
}
\description{
This is the step2 function of the TENETR package.
This function first identifies hm450 DNA methylation probes
that fall within enhancer regions based on enhancer and open chromatin
datasets from the step1 make_external_datasets function as well as
a specified distance from GENCODE v22 transcript TSS.
After identifying enhancer DNA methylation probes, the function classifies
them as methylated/unmethylated or hypermethylated/hypomethylated based on
their differential methylation between the control/normal and
experimental/tumor samples, based on input parameters from the user
}
