% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pairs_z_score.R
\name{get_pairs_z_score}
\alias{get_pairs_z_score}
\title{get_pairs_z_score}
\usage{
get_pairs_z_score(
  TENET_directory,
  hypometh_pairs,
  hypermeth_pairs,
  usecaseonly,
  TF_only,
  hypomethcutoff,
  hypermethcutoff,
  minExp,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step2 results from get_diffmeth_regions function. This function will also create a new step3 folder there containing the results.}

\item{hypometh_pairs}{Set to TRUE/FALSE depending on if you want to calculate z-scores for hypomethylated probes.}

\item{hypermeth_pairs}{Set to TRUE/FALSE depending on if you want to calculate z-scores for hypermethylated probes.}

\item{usecaseonly}{Set to TRUE/FALSE depending on if you want to include the control (normal) samples with the experimental (tumor) samples when calculating hyper/hypomethylated groups and z-scores.}

\item{TF_only}{Set to TRUE/FALSE to determine if you only want to consider genes that are accepted transcription factors in The Human Transcriptoin Factors by Lambert et al (2018) when calculating z-scores.}

\item{hypomethcutoff}{Set a number from 0 to 1 to be the beta-value cutoff for hypomethylated probes.}

\item{hypermethcutoff}{Set a number from 0 to 1 to be the beta-value cutoff for hypermethylated probes.}

\item{minExp}{Sets the minimum number of experimental (tumor) samples to be considered for the hypo/hypermethylated groups.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns tab-delimited ".txt"zscore_all_genes_rda.txt" for each probe of the selected types analyzed, containing the zscore
}
\description{
This is the step3 function of the TENETR package.
This function takes the identified hyper and hypomethylated probes
from step2 get_diffmeth_regions function and calculates a zscore comparing
the mean expression of each gene in groups that are hyper/hypomethylated for
each probe and those that are not, and permutates this across all
hyper/hypomethylated probes
}
