% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_external_datasets.R
\name{make_external_datasets}
\alias{make_external_datasets}
\title{make_external_datasets}
\usage{
make_external_datasets(
  extENH,
  extNDR,
  consensusENH,
  consensusNDR,
  extENH_directory,
  extNDR_directory,
  output_directory,
  core_count
)
}
\arguments{
\item{extENH}{Set TRUE or FALSE depending on if user has enhancer datasets they would like to analyze.}

\item{extNDR}{Set TRUE or FALSE depending on if user has open chromatin datasets they would like to analyze}

\item{consensusENH}{Set TRUE or FALSE depending on if user would like to use consensus enhancer data from TENETR.data for analysis.}

\item{consensusNDR}{Set TRUE or FALSE depending on if user would like to use consensus open chromatin data from TENETR.data for analysis.}

\item{extENH_directory}{Set a path to a directory containing either .bed, .narrowPeak, .broadPeak, or .gappedPeak files with enhancer datasets. This must be supplied and extENH set to TRUE for this analysis.}

\item{extNDR_directory}{Set a path to a directory containing either .bed, .narrowPeak, .broadPeak, or .gappedPeak files with open chromatin datasets. This must be supplied and extNDR set to TRUE for this analysis.}

\item{output_directory}{Set a path to a directory where you want TENETR data to be exported to. Function will create a step1 folder in that directory and subfolders containing function output datasets.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Returns .probelist.txt with the same name as the input files listing the hg38-annotated HM450 DNA methylation probes that fell within the regions specified by each file. These files will be used in downsteam TENETR analyses.
}
\description{
This is the step1 function of the TENETR package.
This function allows users to use bed-like files (see: https://genome.ucsc.edu/FAQ/FAQformat.html#format1)
which contain either enhancer or open chromatin data contained within given directories
or to use pre developed consensus enhancer and open chromatin datasets from many cell/tissue types
from the TENETR.data package, and identify the hg38-annotated HM450 DNA methylation probes
that fall in the regions from each file, then exports .probelist.txt files listing the probes
to a directory specified by the user
}
