% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_links.R
\name{optimize_links}
\alias{optimize_links}
\title{optimize_links}
\usage{
optimize_links(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  adj_pval_cutoff,
  hyper_stringency,
  hypo_stringency,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step4 results from permutate_z_scores function. This function will also create a new step5 folder there containing the results.}

\item{hypermeth_Gplus_analysis}{Set TRUE or FALSE if user wants to optimize hypermeth_Gplus links. Requires hypermeth_analysis from step4 to have been set to TRUE.}

\item{hypermeth_Gminus_analysis}{Set TRUE or FALSE if user wants to optimize hypermeth_Gminus links. Requires hypermeth_analysis from step4 to have been set to TRUE.}

\item{hypometh_Gplus_analysis}{Set TRUE or FALSE if user wants to optimize hypometh_Gplus links. Requires hypometh_analysis from step4 to have been set to TRUE.}

\item{hypometh_Gminus_analysis}{Set TRUE or FALSE if user wants to optimize hypometh_Gminus links. Requires hypometh_analysis from step4 to have been set to TRUE.}

\item{adj_pval_cutoff}{Set p-value for BH-corrected Wilcoxon p-values in comparison of gene expression values between hyper/hypomethylated tumor/experimental samples and the control/normal samples..}

\item{hyper_stringency}{Set a number from 0 to 1 to be the beta-value cutoff to optimize for hypermeth links with maximum methylation values above the cutoff}

\item{hypo_stringency}{Set a number from 0 to 1 to be the beta-value cutoff to optimize for hypometh links with minimum methylation values below the cutoff}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns tab-delimited "sig_link_zscores_perm_optimized.txt" files for hypo/hyper Gplus/Gminus probe-gene links, similar to step4, but only with the optimized links.
}
\description{
This is the step5 function of the TENETR package.
This function takes the calculated permutated p-values for the hyper/hypomethylated
Gplus or Gminus probe-gene quadrants, and selects optimized links based on relative
expression of the gene in hyper/hypomethylated experimental/tumor samples and the control/normal samples
based on their means and in a wilcoxon test, checks that the hyper/hypometh samples for that given
gene-probe link also show appropriately higher/lower expression of the linked gene in a number of
experimental/tumor samples greater than the minExp number specified in step2 get_diffmeth_regions function
and have maximum/minimum methylation above/below the hyper_stringency and hypo_stringency cutoffs.
}
