% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutate_z_scores.R
\name{permutate_z_scores}
\alias{permutate_z_scores}
\title{permutate_z_scores}
\usage{
permutate_z_scores(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step3 results from get_analysis_z_score function. This function will also create a new step4 folder there containing the results.}

\item{hypermeth_Gplus_analysis}{Set TRUE or FALSE if user wants to permutate on hypermeth_Gplus links. Requires hypermeth_analysis from step3 to have been set to TRUE.}

\item{hypermeth_Gminus_analysis}{Set TRUE or FALSE if user wants to permutate on hypermeth_Gminus links. Requires hypermeth_analysis from step3 to have been set to TRUE.}

\item{hypometh_Gplus_analysis}{Set TRUE or FALSE if user wants to permutate on hypometh_Gplus links. Requires hypometh_analysis from step3 to have been set to TRUE.}

\item{hypometh_Gminus_analysis}{Set TRUE or FALSE if user wants to permutate on hypometh_Gminus links. Requires hypometh_analysis from step3 to have been set to TRUE.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns tab-delimited tab-delimited "sig_link_zscores_perm.txt" files for hypo/hyper Gplus/Gminus probe-gene links, similar to step3, but with the permutated p-value for each link.
}
\description{
This is the step4 function of the TENETR package.
This function takes the calculated z-scores for the hyper/hypomethylated
Gplus or Gminus probe-gene quadrants, and does a pseudo-permutation,
calculating individual permutated p-values for the probe-gene z-score
based on their relative rank amongst all probes to a given gene.
}
