% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_make_external_datasets_test.R
\name{step1_make_external_datasets_test}
\alias{step1_make_external_datasets_test}
\title{step1_make_external_datasets_test}
\usage{
step1_make_external_datasets_test(
  output_directory,
  DNA_methylation_manifest = "HM450",
  ext_HM = FALSE,
  ext_NDR = FALSE,
  consensus_ENH = TRUE,
  consensus_NDR = TRUE,
  core_count
)
}
\arguments{
\item{output_directory}{Set a path to a directory, which will be created by this function, if it doesn't already exist, where you want the .probelist.txt files created by this function to be exported to. This directory will be the 'TENET_directory' utilized in later functions and a 'step1' subdirectory will be created in that directory, with further subdirectories containing the output .probelist.txt files made by this function.}

\item{DNA_methylation_manifest}{Set to 'HM27', 'HM450', or 'EPIC' depending on the DNA methylation array of interest for the user's data. hg38 array annotations come from https://zwdzwd.github.io/InfiniumAnnotation. Defaults to 'HM450'.}

\item{ext_HM}{If the user has their own histone modification datasets, set this as a path to a directory containing either .bed, .narrowPeak, .broadPeak, or .gappedPeak files with the user's own histone modification datasets. Otherwise, set to FALSE. Defaults to FALSE.}

\item{ext_NDR}{If the user has their own open chromatin, or NDR, datasets, set this as a path to a directory containing either .bed, .narrowPeak, .broadPeak, or .gappedPeak files with the user's own open chromatin datasets. Otherwise, set to FALSE. Defaults to FALSE.}

\item{consensus_ENH}{Set TRUE or FALSE depending on if user would like to use the consensus enhancer histone modification data from TENETR.data for analysis. Defaults to TRUE.}

\item{consensus_NDR}{Set TRUE or FALSE depending on if user would like to use the consensus open chromatin data from TENETR.data for analysis. Defaults to TRUE.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Returns .probelist.txt fils with the same name as the input files listing the hg38-annotated DNA methylation probes that fell within the regions specified by each file. These files will be used in later TENETR analyses.
}
\description{
This is the step1 function of the TENETR package.
This function allows users to use bed-like files see: https://genome.ucsc.edu/FAQ/FAQformat.html#format1
which contain either histone modification or open chromatin data contained within directories
specified by the user, or to use pre developed consensus enhancer and open chromatin
datasets from many cell/tissue types contained in the TENETR.data package,
in order to identify the hg38-annotated DNA methylation probes of the
user-specified array that fall in the peaks/regions specified by each file,
then saves .probelist.txt files listing the probes to a directory specified
by the user, which will also act as the 'TENET_directory' for future functions
in the TENETR package.
}
