% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_get_diffmeth_regions_test.R
\name{step2_get_diffmeth_regions_test}
\alias{step2_get_diffmeth_regions_test}
\title{step2_get_diffmeth_regions_test}
\usage{
step2_get_diffmeth_regions_test(
  TENET_directory,
  methylation_expression_dataset,
  DNA_methylation_manifest = "HM450",
  assess_promoter = FALSE,
  TSS_dist = 1500,
  use_ext_HM = FALSE,
  use_ext_NDR = FALSE,
  use_consensus_ENH = TRUE,
  use_consensus_NDR = TRUE,
  methcutoff = FALSE,
  hypomethcutoff = FALSE,
  unmethcutoff = FALSE,
  hypermethcutoff = FALSE,
  min_experimental_count,
  purity_directory = FALSE
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step1' subdirectory and results created by the step1_make_external_datasets function. This function will also create a new 'step2' subdirectory there containing the results of this function.}

\item{methylation_expression_dataset}{Set a path to the an .rda file containing methylation and expression data, possibly created using the TCGA_downloader function. This dataset should contain 'expDataN' and 'expDataT' expression datasets, as well as 'metDataN' and 'metDataT' methylation datasets, and optionally a 'clinical' dataset with sample clinical information.}

\item{DNA_methylation_manifest}{Set to 'HM27', 'HM450', or 'EPIC' depending on the DNA methylation array of interest for the user's data. hg38 array annotations come from https://zwdzwd.github.io/InfiniumAnnotation. Defaults to 'HM450'.}

\item{assess_promoter}{Set TRUE or FALSE depending on if you want to identify DNA methylation probes that mark promoter regions or distal enhancer regions, respectively. Defaults to FALSE.}

\item{TSS_dist}{Set a positive integer to be the buffer in base pairs from GENCODE v22-annotated transcription start sites for DNA methylation probes to be considered promoter probes. Probes outside of the TSS_dist from annotated transcription start sites will be considered enhancer probes. Defaults to 1500.}

\item{use_ext_HM}{Set TRUE or FALSE depending on if you want to use DNA methylation probes in user-supplied histone modification datasets from the step1_make_external_datasets function. These probes will be combined with probes found in the consensus enhancer regions if 'use_consensus_ENH' is also set to TRUE. These will be overlapped with probes from nucleosome depleted regions if either 'use_ext_NDR' or 'use_consensus_NDR' are set to TRUE. Setting to TRUE requires that the user have run the step1_make_external_datasets function on user-supplied histone modification datasets previously. Defaults to FALSE.}

\item{use_ext_NDR}{Set TRUE or FALSE depending on if you want to use DNA methylation probes in user-supplied nucleosome depeleted regions datasets from the step1_make_external_datasets function. These probes will be combined with probes found in the consensus nucleosome depeleted regions if 'use_consensus_NDR' is also set to TRUE. These will be overlapped with probes from regions with histone modifications or enhancer regions if either 'use_ext_HM' or 'use_consensus_ENH' are set to TRUE. Setting to TRUE requires that the user have run the step1_make_external_datasets function on user-supplied nucleosome depleted regions datasets previously. Defaults to FALSE.}

\item{use_consensus_ENH}{Set TRUE or FALSE depending on if you want to use DNA methylation probes in consensus enhancer regions from the TENETR.data package and step1_make_external_datasets. These probes will be combined with probes found in user-supplied histone modification regions if 'use_ext_HM' is also set to TRUE. These will be overlapped with probes in nucleosome depleted regions if either 'use_ext_NDR' or 'use_consensus_NDR' are set to TRUE. Setting to TRUE requires that the user have run step1_make_external_datasets on consensus enhancer datasets from the TENETR.data package previously. Defaults to TRUE.}

\item{use_consensus_NDR}{Set TRUE or FALSE depending on if you want to use DNA methylation probes in consensus nucleosome depeleted regions from the TENETR.data package and step1_make_external_datasets. These probes will be combined with probes found in user-supplied nucleosome depeleted regions if 'use_ext_NDR' is also set to TRUE. These will be overlapped with probes from regions with histone modifications or enhancer regions if either 'use_ext_HM' or 'use_consensus_ENH' are set to TRUE. Setting to TRUE requires that the user have run step1_make_external_datasets on consensus nucleosome depleted regions datasets from the TENETR.data package previously. Defaults to TRUE.}

\item{methcutoff}{Set a number from 0 to 1 to be the beta-value cutoff for methylated probes. Otherwise, will use algorithm to set optimal cutoff value. Defaults to using algorithm if no input is provided.}

\item{hypomethcutoff}{Set a number from 0 to 1 to be the beta-value cutoff for hypomethylated probes. Should be set lower than the methcutoff. Otherwise, will use algorithm to set optimal cutoff value. Defaults to using algorithm if no input is provided.}

\item{unmethcutoff}{Set a number from 0 to 1 to be the beta-value cutoff for unmethylated probes. Otherwise, will use algorithm to set optimal cutoff value. Defaults to using algorithm if no input is provided.}

\item{hypermethcutoff}{Set a number from 0 to 1 to be the beta-value cutoff for hypermethylated probes. Should be set higher than the unmethcutoff. Otherwise, will use algorithm to set optimal cutoff value. Defaults to using algorithm if no input is provided.}

\item{min_experimental_count}{Set a positive integer to be the minimum number of experimental/tumor samples to be considered for the hypo/hypermethylated groups. Should be less than the total number of experimental/tumor groups.}

\item{purity_directory}{If this user has their own datasets containing data for possible cell types that may contaminate the data, as .rda files containing a data matrix with DNA methylation values, set this as the path to the directory containing the purity .rda files. Otherwise set to FALSE. Defaults to FALSE.}
}
\value{
Returns up to four objects, a .rda file with matrices of methylation and expression data for control/normal and experimental/tumor sample, lists of the four types of identified DNA methylation probes marking the regulatory elements of interest, the clinical data if provided, and the user-set parameters for consistency in downstream analyses. Also outputs a .txt file containing metrics on the number of probes found in the different analysis categories, and the specified cutoff values, a .pdf file showing where the cutoffs were positioned on the control and experimental sample mean density curves, and finally a .txt file showing which input HM and NDR datasets from step1 the probes happened to fall within.
}
\description{
This is the step2 function of the TENETR package.
This function identifies DNA methylation probes that mark putative regulatory
elements, including enhancer and promoter regions. These are probes that lie within
regions with specific histone modifications and open chromatin regions,
from the step1_make_external_datasets function, and which are located
a user-specified distance relative to GENCODE v22 transcript transcription start sites.
After identifying DNA methylation probes representing the the specified regulatory elements,
the function classifies the probes as methylated, unmethylated, hypermethylated,
or hypomethylated based on their differential methylation between the control/normal
and experimental/tumor samples supplied by the user, defined by the cutoff values
also specified by the user.
}
