% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_permutate_z_scores.R
\name{step4_permutate_z_scores}
\alias{step4_permutate_z_scores}
\title{step4_permutate_z_scores}
\usage{
step4_permutate_z_scores(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step3' subdirectory and results created by the step3_get_analysis_z_scores function. This function will also create a new 'step4' subdirectory there containing the results of this function.}

\item{hypermeth_Gplus_analysis}{Set TRUE or FALSE if user wants to permutate on hypermeth_Gplus probe-gene links. Requires hypermeth_analysis from step3 to have been set to TRUE.}

\item{hypermeth_Gminus_analysis}{Set TRUE or FALSE if user wants to permutate on hypermeth_Gminus probe-gene links. Requires hypermeth_analysis from step3 to have been set to TRUE.}

\item{hypometh_Gplus_analysis}{Set TRUE or FALSE if user wants to permutate on hypometh_Gplus probe-gene links. Requires hypometh_analysis from step3 to have been set to TRUE.}

\item{hypometh_Gminus_analysis}{Set TRUE or FALSE if user wants to permutate on hypometh_Gminus probe-gene links. Requires hypometh_analysis from step3 to have been set to TRUE.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns tab-delimited "sig_link_zscores_perm.txt" files for hypo/hyper Gplus/Gminus probe-gene links, similar to step3, but with the permutated p-value for each link rather than z-score.
}
\description{
This is the step4 function of the TENETR package.
This function takes the calculated z-scores for the hyper or hypomethylated
Gplus or Gminus probe-gene links, depending on user input, and performs a
permutation analysis, calculating individual permutated p-values for each
probe-gene z-score from the previous step based on their relative rank amongst
the link z-scores for all enhancer DNA methylation probes to a given gene.
}
