% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5_optimize_links.R
\name{step5_optimize_links}
\alias{step5_optimize_links}
\title{step5_optimize_links}
\usage{
step5_optimize_links(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  adj_pval_cutoff,
  hyper_stringency,
  hypo_stringency,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step4' subdirectory and results created by the step4_permutate_z_scores function. This function will also create a new 'step5' subdirectory there containing the results of this function.}

\item{hypermeth_Gplus_analysis}{Set TRUE or FALSE if user wants to optimize hypermeth_Gplus links. Requires hypermeth_analysis from step4 to have been set to TRUE.}

\item{hypermeth_Gminus_analysis}{Set TRUE or FALSE if user wants to optimize hypermeth_Gminus links. Requires hypermeth_analysis from step4 to have been set to TRUE.}

\item{hypometh_Gplus_analysis}{Set TRUE or FALSE if user wants to optimize hypometh_Gplus links. Requires hypometh_analysis from step4 to have been set to TRUE.}

\item{hypometh_Gminus_analysis}{Set TRUE or FALSE if user wants to optimize hypometh_Gminus links. Requires hypometh_analysis from step4 to have been set to TRUE.}

\item{adj_pval_cutoff}{Set p-value for Benjamini-Hochberg corrected Wilcoxon p-values in comparison of gene expression values between hyper/hypomethylated tumor/experimental samples and the control/normal samples..}

\item{hyper_stringency}{Set a number from 0 to 1 to be the beta-value cutoff to optimize for hypermeth links with methylation values above the cutoff}

\item{hypo_stringency}{Set a number from 0 to 1 to be the beta-value cutoff to optimize for hypometh links with methylation values below the cutoff}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns tab-delimited "sig_link_zscores_perm_optimized.txt" files for hypo/hyper Gplus/Gminus probe-gene links, similar to step4, but only with the optimized links.
}
\description{
This is the step5 function of the TENETR package.
This function takes the calculated permutated p-values for the hyper or hypomethylated
Gplus or Gminus probe-gene links, and selects optimized links based on relative
expression of the gene in hyper or hypomethylated experimental/tumor samples
as well as the control/normal samples based on their means and, using a Wilcoxon test,
checks that the hyper or hypomethylated samples for that given gene-probe link
also show appropriately higher/lower expression of the linked gene in a number of
experimental/tumor samples greater than the minExp number specified in the
step2_get_diffmeth_regions function and which have maximum/minimum methylation
above/below the hyper_stringency and hypo_stringency cutoff values selected.
}
