% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step6_probe_per_gene_tabulation.R
\name{step6_probe_per_gene_tabulation}
\alias{step6_probe_per_gene_tabulation}
\title{step6_probe_per_gene_tabulation}
\usage{
step6_probe_per_gene_tabulation(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step5' subdirectory and results created by the step5_optimize_links function. This function will also create a new 'step6' subdirectory there containing the results of this function.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to calculate total links by gene/TF for hypermeth probes with G+ links.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to calculate total links by gene/TF for hypermeth probes with G- links.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to calculate total links by gene/TF for hypometh probes with G+ links.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to calculate total links by gene/TF for hypometh probes with G- links.}
}
\value{
Currently returns tab-delimited "_freq.txt" files that include all genes, or annotated TFs only, organized by decreasing number of linked probes of the given analysis type.
}
\description{
This is the step6 function of the TENETR package.
This function takes the final optimized enhancer DNA methylation probe to gene
links found from step5_optimize_links and tabulates how many of these links are
found, both for each gene in the analysis as well as only accepted transcription factor
genes from "The Human Transcription Factors" by Lambert et al. 2018.
This tabulation is done separately for each of the four hypo or hypermethylated
Gplus or Gminus analysis quadrants, as selected by the user.
}
