% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7_linked_probe_motif_searching.R
\name{step7_linked_probe_motif_searching}
\alias{step7_linked_probe_motif_searching}
\title{step7_linked_probe_motif_searching}
\usage{
step7_linked_probe_motif_searching(
  TENET_directory,
  TF_gene,
  DNA_methylation_manifest = "HM450",
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  motif_PCM_PWM,
  distance_from_probes = 100,
  matchPWM_min_score = "75\%",
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step6' subdirectory and results created by the step6_probe_per_gene_tabulation function. This function will also create a new 'step7' subdirectory there, if not already created, with a further subdirectories for each of the four analysis types selected, ending with '_survival' containing the results of this function.}

\item{TF_gene}{Specify a transcription factor gene by its name or ENSG ID to perform motif searching for.}

\item{DNA_methylation_manifest}{Set to 'HM27', 'HM450', or 'EPIC' depending on the DNA methylation array of interest for the user's data. hg38 array annotations come from https://zwdzwd.github.io/InfiniumAnnotation. Defaults to 'HM450'.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to do motif searching in the vicintiy of hypermeth probes with G+ links to the TF of interest, if specified.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to do motif searching in the vicintiy of hyometh probes with G- links to the TF of interest, if specified.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to do motif searching in the vicintiy of hypometh probes with G+ links to the TF of interest, if specified.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to do motif searching in the vicintiy of hypermeth probes with G- links to the TF of interest, if specified.}

\item{motif_PCM_PWM}{Specify a PCM or PWM in the form of a 4xN matrix with the motif for the specified TF_gene. If a PCM or PWM is not available, the user can select one by querying the MotifDb package database. See examples below for more info.}

\item{distance_from_probes}{Set a positive integer in base pairs to be the distance added to the beginning and from end of the the linked enhancer probes to perform motif searching on. Defaults to 100.}

\item{matchPWM_min_score}{Set a positive integer, from 1 to 100, to be the min.score passed to the matchPWM argument for motif searching. See ?matchPWM from the Biostrings package for more info. Defaults to 75\%.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Returns a .pdf file containing a seqLogo for the specified PCM or PWM, as well as two tables. The first contains information about each motif found in the vicintiy of the linked DNA methylation probes for the specified TF_gene at the specified matchPWM_min_score. The second lists each probe linked to the TF_gene for the analysis types specified, and the number of motif occurrences found in the vicinity of each of those probes.
}
\description{
This is a step7 function of the TENETR package.
This function takes a binding motif, supplied by the user, for a
TF with linked probes also specified by the user, and for each of those
linked probes, it identifies if a motif for that TF is found in
a user-specified vicinity of each probe linked to the TF.
}
\examples{
# Show available motifs for example TF FOXM1:
names(query(MotifDb,"FOXM1")
# Once you've selected a PWM for use, you can specify it using the following:
motif_PCM_PWM= query(MotifDb, motif_name)[[3]]
}
