% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7_selected_probes_simple_scatterplots.R
\name{step7_selected_probes_simple_scatterplots}
\alias{step7_selected_probes_simple_scatterplots}
\title{step7_selected_probes_simple_scatterplots}
\usage{
step7_selected_probes_simple_scatterplots(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  probe_list,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step6' subdirectory and results created by the step6_probe_per_gene_tabulation function. This function will also create a new 'step7' subdirectory there, if not already created, with an additional a subdirectory 'selected_probes_simple_scatterplots' containing the results of this function.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create scatterplots for the user-supplied probes with all genes the probes may have hypermeth Gplus links to.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to create scatterplots for the user-supplied probes with all genes the probes may have hypermeth Gminus links to.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create scatterplots for the user-supplied probes with all genes the probes may have hypometh Gplus links to.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to create scatterplots for the user-supplied probes with all genes the probes may have hypometh Gminus links to.}

\item{probe_list}{Supply a vector of DNA methylation probe names for which scatterplots of those probes with expression of any linked genes of the specified analysis types above will be generated.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns .pdf files with scatterplots showing the expression of any genes linked to the specified probes on the x-axis and the methylation level of the probes on the y-axis.
}
\description{
This is a step7 function of the TENETR package.
This function takes a custom list of DNA methylation probes specified by the user
and generates scatterplots displaying the expression level of the genes
linked to these probes for each of the four hypo or hypermethylated
Gplus or Gminus analysis quadrants, as selected by the user.
}
