% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7_states_for_links.R
\name{step7_states_for_links}
\alias{step7_states_for_links}
\title{step7_states_for_links}
\usage{
step7_states_for_links(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step6' subdirectory and results created by the step6_probe_per_gene_tabulation function. This function will also create a new 'step7' subdirectory there, if not already created, with a further subdirectories for each of the four analysis types selected, ending with '_states_for_links' containing the results of this function.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create .tsv files showing which experimental/tumor samples harbor a given hypermeth Gplus probe-gene link.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create .tsv files showing which experimental/tumor samples harbor a given hypermeth Gminus probe-gene link.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create .tsv files showing which experimental/tumor samples harbor a given hypometh Gplus probe-gene link.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create .tsv files showing which experimental/tumor samples harbor a given hypometh Gminus probe-gene link.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Exports .tsv files for each of the specified analysis types with experimental/tumor samples in the columns and each probe-gene link for that analysis type in the rows, with a 1 indicating the sample is positive for that link, and a 0 if that sample isn't.
}
\description{
This is a step7 function of the TENETR package.
This function generates .tsv files with information for each of the experimental/tumor,
samples and each probe-gene link from each of the four hypo or hypermethylated
Gplus or Gminus analysis quadrants, as selected by the user, on if a given sample
is said "harbor" each link, depending on if the methylation of the given sample
for the probe in the link is above or below the hyper or hypomethcutoff defined
from the step2_get_diffmeth_regions function, and the expression of gene in the link
is significantly greater than, or less than, the mean expression in the normal/control
samples.
}
