% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7_top_genes_met_heatmaps_test.R
\name{step7_top_genes_met_heatmaps_test}
\alias{step7_top_genes_met_heatmaps_test}
\title{step7_top_genes_met_heatmaps_test}
\usage{
step7_top_genes_met_heatmaps_test(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  top_gene_number,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step6' subdirectory and results created by the step6_probe_per_gene_tabulation function. This function will also create a new 'step7' subdirectory there, if not already created, with further subdirectories for each of the four analysis types selected, ending with '_met_heatmaps' containing the results of this function.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create heatmaps showing DNA methylation levels of enhancer probes linked to the top genes/TFs with the most hypermeth probes with G+ links.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create heatmaps showing DNA methylation levels of enhancer probes linked the top genes/TFs with the most hypermeth probes with G- links.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create heatmaps showing DNA methylation levels of enhancer probes linked the top genes/TFs with the most hypometh probes with G+ links.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create heatmaps showing DNA methylation levels of enhancer probes linked the top genes/TFs with the most hypometh probes with G- links.}

\item{top_gene_number}{Specify a number of the top genes/TFs based on the most linked enhancer probes of a given analysis type to generate heatmaps for with their linked enhancer probes methylation levels.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns .pdf files with the heatmaps showing the DNA methylation levels of all probes of a specific analysis type linked to those genes, as well as the expression of said genes in the column labels
}
\description{
This is a step7 function of the TENETR package.
This function takes the top genes/TFs by number of linked probes identified from
the step6_probe_per_gene_tabulation function up to the number as specified by the user
and generates heatmaps showing the methylation level of the enhancer
DNA methylation probes linked to those genes of the specified analysis types.
}
