% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7_top_genes_simple_scatterplots_test.R
\name{step7_top_genes_simple_scatterplots_test}
\alias{step7_top_genes_simple_scatterplots_test}
\title{step7_top_genes_simple_scatterplots_test}
\usage{
step7_top_genes_simple_scatterplots_test(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  top_gene_number,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step6' subdirectory and results created by the step6_probe_per_gene_tabulation function. This function will also create a new 'step7' subdirectory there, if not already created, with further subdirectories for each of the four analysis types selected, ending with '_simple_scatterplots' containing the results of this function.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create scatterplots for the top genes/TRs by most hypermeth probes with G+ links and each of their linked probes.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create scatterplots for the top genes/TRs by most hypermeth probes with G- links and each of their linked probes.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create scatterplots for the top genes/TRs by most hypometh probes with G+ links and each of their linked probes.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create scatterplots for the top genes/TRs by most hypometh probes with G- links and each of their linked probes.}

\item{top_gene_number}{Specify a number of the top genes/TFs based on the most linked enhancer probes of a given analysis type to generate scatterplots for showing expression of the genes and methylation of each of their linked enhancer probes.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns .pdf files with scatterplots showing the expression of the genes of interest on the x-axis and the methylation of the linked probes on the y-axis.
}
\description{
This is a step7 function of the TENETR package.
This function takes the top genes/TFs by number of linked probes identified from
the step6_probe_per_gene_tabulation function up to the number as specified by the user
and generates scatterplots displaying the expression level of each of these genes
and the methylation level of each enhancer probe linked to them for each of the
four hypo or hypermethylated Gplus or Gminus analysis quadrants, as selected by the user.
}
