% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7_top_genes_UCSC_bed_files.R
\name{step7_top_genes_UCSC_bed_files}
\alias{step7_top_genes_UCSC_bed_files}
\title{step7_top_genes_UCSC_bed_files}
\usage{
step7_top_genes_UCSC_bed_files(
  TENET_directory,
  DNA_methylation_manifest = "HM450",
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  top_gene_number,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step6' subdirectory and results created by the step6_probe_per_gene_tabulation function. This function will also create a new 'step7' subdirectory there, if not already created, with further subdirectories for each of the four analysis types selected, ending with '_UCSC_bed_files' containing the results of this function.}

\item{DNA_methylation_manifest}{Set to 'HM27', 'HM450', or 'EPIC' depending on the DNA methylation array of interest for the user's data. hg38 array annotations come from https://zwdzwd.github.io/InfiniumAnnotation. Defaults to 'HM450'.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create interact .bed files that can be uploaded to the UCSC genome browser showing links between the top genes/TFs by most hypermeth probes with G+ links, and said linked probes.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create interact .bed files that can be uploaded to the UCSC genome browser showing links between the top genes/TFs by most hypermeth probes with G- links, and said linked probes.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create interact .bed files that can be uploaded to the UCSC genome browser showing links between the top genes/TFs by most hypometh probes with G+ links, and said linked probes.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create interact .bed files that can be uploaded to the UCSC genome browser showing links between the top genes/TFs by most hypometh probes with G- links, and said linked probes.}

\item{top_gene_number}{Specify a number of the top genes/TFs based on the most linked enhancer probes to generate interact bed files for showing the links between the genes and each of their linked enhancer probes.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Exports .bed formatted interact files to upload to the UCSC genome browser. These files display the interactions between the top genes/TFs and their linked enhancer probes for the given analysis types.
}
\description{
This is a step7 function of the TENETR package.
This function takes the top genes/TFs by number of linked probes identified from
the step6_probe_per_gene_tabulation function up to the number as specified by the user
and generates interact files see: https://genome.ucsc.edu/goldenPath/help/interact.html
that can be uploaded to the UCSC genome browser and visualize the
links between each of the top specified genes/TFs and the enhancer probes linked
to them for each of the four hypo or hypermethylated Gplus or Gminus analysis quadrants,
as selected by the user.
}
