% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7_top_genes_user_peak_overlap.R
\name{step7_top_genes_user_peak_overlap}
\alias{step7_top_genes_user_peak_overlap}
\title{step7_top_genes_user_peak_overlap}
\usage{
step7_top_genes_user_peak_overlap(
  TENET_directory,
  ext_peaks_directory,
  DNA_methylation_manifest = "HM450",
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  top_gene_number,
  distance_from_probes = 100,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the TENET directory containing the 'step6' subdirectory and results created by the step6_probe_per_gene_tabulation function. This function will also create a new 'step7' subdirectory there, if not already created, with a further subdirectories for each of the four analysis types selected, ending with '_user_peak_overlap' containing the results of this function.}

\item{ext_peaks_directory}{If the user has their own datasets with peaks for factors interest, set this as a path to a directory containing either .bed, .narrowPeak, .broadPeak, or .gappedPeak files with these peaks.}

\item{DNA_methylation_manifest}{Set to 'HM27', 'HM450', or 'EPIC' depending on the DNA methylation array of interest for the user's data. hg38 array annotations come from https://zwdzwd.github.io/InfiniumAnnotation. Defaults to 'HM450'.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create survival plots for the top genes/TFs by most hypermeth probes with G+ links, and these linked DNA methyation probes if specified.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create survival plots for the top genes/TFs by most hypermeth probes with G- links, and these linked DNA methyation probes if specified.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create survival plots for the top genes/TFs by most hypometh probes with G+ links, as well as their linked DNA methyation probes if specified.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create survival plots for the top genes/TFs by most hypometh probes with G- links, as well as their linked DNA methyation probes if specified.}

\item{top_gene_number}{Specify a number of the top genes/TFs based on the most linked enhancer probes to generate files showing overlap with user peak files for the enhancer probes linked to those genes.}

\item{distance_from_probes}{Set a positive integer in base pairs to be the distance added to the beginning and from end of the the linked enhancer probes to perform peak searching on. Defaults to 100.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Returns a .txt file for each of the top genes/TFs showing which of the user-specified peak files the enhancer DNA methylation probes linked to each of the genes overlap with.
}
\description{
This is a step7 function of the TENETR package.
This function takes the top genes/TFs by number of linked probes identified from
the step6_probe_per_gene_tabulation function up to the number as specified by the user
and identifies if the probes linked to those TFs are found in peaks for factors of
interest, supplied by the user in the form of .bed, .narrowPeak, .gappedPeak,
or .broadPeak files in a user-specified directory.
}
