% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tr_met_heatmaps.R
\docType{data}
\name{TENET_directory}
\alias{TENET_directory}
\title{top_tr_met_heatmaps}
\format{
An object of class \code{character} of length 1.
}
\usage{
TENET_directory
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step6 results from the top_tr_tabulation function. This function will also create a new step7 folder there if it has not been created, with a subdirectory with 'met_heatmaps' containing the results.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create heatmaps showing DNA methylation levels of probes linked to the top TRs with the most hypermeth probes with G+ links.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create heatmaps showing DNA methylation levels of probes linked the top TRs with the most hypermeth probes with G- links.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create heatmaps showing DNA methylation levels of probes linked the top TRs with the most hypometh probes with G+ links.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create heatmaps showing DNA methylation levels of probes linked the top TRs with the most hypometh probes with G- links.}

\item{top_gene_number}{Specify a number to generate heatmaps for the probes linked to that number of the top genes/TFs based on the most linked enhancer probes.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns .pdf files with the heatmaps showing the DNA methylation levels of all probes of a specific analysis type linked to those genes, as well as the expression of said genes across the samples in the column labels
}
\description{
This is a step7 function of the TENETR package.
This function takes the top genes/TRs by number of linked probes identified from
the step6 top_tr_tabulation function up to the number as specified by the user
and generates heatmaps showing the methylation level of the enhancer
DNA methylation probes linked to those genes.
}
\keyword{datasets}
