% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tr_circos.R
\name{top_tr_circos}
\alias{top_tr_circos}
\title{top_tr_ciros}
\usage{
top_tr_circos(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  top_gene_number,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step6 results from the top_tr_tabulation function. This function will also create a new step7 folder there if it has not been created, with a subdirectory called histogram containing the results.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create histograms for the top TRs by most hypermeth probes with G+ links.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypermeth probes with G- links.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypometh probes with G+ links.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypometh probes with G- links.}

\item{top_gene_number}{Specify a number to generate circos plots for that many of the top genes and TFs based on the most linked enhancer probes.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns .pdf files with the histograms showing the number of TRs linked to a given number of probes of the given analysis type.
}
\description{
This is a step7 function of the TENETR package.
This function takes the top TRs by number of linked probes identified from
the step6 top_tr_tabulation function up to the number as specified by the user
and generates circos plots for each gene showing the genomic links between
each gene and each probe linked to it.
}
