% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tr_histogram.R
\name{top_tr_histogram}
\alias{top_tr_histogram}
\title{top_tr_histogram}
\usage{
top_tr_histogram(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step6 results from the top_tr_tabulation function. This function will also create a new step7 folder there if it has not been created, with a subdirectory called histogram containing the results.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create histograms for the top TRs by most hypermeth probes with G+ links.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypermeth probes with G- links.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypometh probes with G+ links.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypometh probes with G- links.}
}
\value{
Currently returns .pdf files with the histograms showing the number of TRs linked to a given number of probes of the given analysis type.
}
\description{
This is a step7 function of the TENETR package.
This function takes the top TRs by number of linked probes identified from
the step6 top_tr_tabulation function and generates heatmaps showing the
number of all genes and TF-only genes with a given number of enhancer
DNA methylation probes of a given type linked to them.
}
