% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tr_simple_scatterplots.R
\name{top_tr_simple_scatterplots}
\alias{top_tr_simple_scatterplots}
\title{top_tr_simple_scatterplots}
\usage{
top_tr_simple_scatterplots(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  top_gene_number,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step6 results from the top_tr_tabulation function. This function will also create a new step7 folder there if it has not been created, with a subdirectory with 'scatterplots' containing subfolders the results for the top genes and top TFs separately.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create scatterplots for the top genes/TRs by most hypermeth probes with G+ links and each of their linked probes.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create scatterplots for the top genes/TRs by most hypermeth probes with G- links and each of their linked probes.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create scatterplots for the top genes/TRs by most hypometh probes with G+ links and each of their linked probes.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create scatterplots for the top genes/TRs by most hypometh probes with G- links and each of their linked probes.}

\item{top_gene_number}{Specify a number to generate scatterplots for that many of the top genes/TFs based on the most linked enhancer probes.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns .pdf files with scatterplots showing the expression of the genes of interest on the x-axis and the methylation of the linked probes on the y-axis.
}
\description{
This is a step7 function of the TENETR package.
This function takes the top genes/TRs by number of linked probes identified from
the step6 top_tr_tabulation function up to the number as specified by the user
and generates scatterplots displaying the expression level of each of these genes
and the methylation level of each enhancer probe linked to them.
}
