% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tr_simple_scatterplots_selected_probes.R
\name{top_tr_simple_scatterplots_selected_probes}
\alias{top_tr_simple_scatterplots_selected_probes}
\title{top_tr_simple_scatterplots_selected_probes}
\usage{
top_tr_simple_scatterplots_selected_probes(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  probe_list,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step6 results from the top_tr_tabulation function. This function will also create a new step7 folder there if it has not been created, with a subdirectory with 'simple_scatterplots' containing the results.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create histograms for the top TRs by most hypermeth probes with G+ links.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypermeth probes with G- links.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypometh probes with G+ links.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypometh probes with G- links.1}

\item{probe_list}{Supply a vector of probe names for which scatterplots for those probs with expression of any linked genes of the specified analysis types above will be generated.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns .pdf files with scatterplots showing the expression of the genes linked to the specified probes on the x-axis and the methylation the probes on the y-axis.
}
\description{
This is a step7 function of the TENETR package.
This function takes a custom list of probes specified by the user
and generates scatterplots displaying the expression level of the genes
linked to these probes in the specified analysis types, i.e.e hyper/hypo
Gplus/Gminus
}
