% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tr_states_for_links.R
\name{top_tr_states_for_links}
\alias{top_tr_states_for_links}
\title{top_tr_states_for_links}
\usage{
top_tr_states_for_links(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step6 results from the top_tr_tabulation function. This function will also create a new step7 folder there if it has not been created, with subdirectories for each of the four analysis types to contain the link files.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create .tsv files showing which tumor samples harbor a given probe-gene link of the given analysis type.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create .tsv files showing which tumor samples harbor a given probe-gene link of the given analysis type.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create .tsv files showing which tumor samples harbor a given probe-gene link of the given analysis type.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create .tsv files showing which tumor samples harbor a given probe-gene link of the given analysis type.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Exports .tsv files for each of the specified analysis types with tumor samples in the y-axis and each probe-gene link for that analysis type in the x-axis, with a 1 indicating the sample is positive for that link, and a 0 if that sample isn't.
}
\description{
This is a step7 function of the TENETR package.
This function takes the top genes/TFs by number of linked probes identified from
the step6 top_tr_tabulation function up to the number as specified by the user
and generates .tsv files with information for each of the experimental, or tumor,
samples for each probe to gene link for the analysis types selected
}
