% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tr_tabulation.R
\name{top_tr_tabulation}
\alias{top_tr_tabulation}
\title{top_tr_tabulation}
\usage{
top_tr_tabulation(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step5 results from the optimize_links function. This function will also create a new step6 folder there containing the results.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to calculate the top_tfs for hypermeth probes with G+ links.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to calculate the top_tfs for hypermeth probes with G- links.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to calculate the top_tfs for hypometh probes with G+ links.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to calculate the top_tfs for hypometh probes with G- links.}
}
\value{
Currently returns tab-delimited "sig_link_zscores.txt" files for hypo/hyper Gplus/Gminus probe-gene links, as well as individual "zscores.txt" files named after each gene in the hypo/hyper analysis with zscores for that gene to all probes from that analysis type.
}
\description{
This is the step6 function of the TENETR package.
This function takes the identified hyper and hypomethylated probes
from step2 get_diffmeth_regions function and calculates a zscore comparing
the mean expression of each gene in groups that are hyper/hypomethylated for
each probe, according to hypermeth and hypometh and those that are not, and permutates this across all
hyper/hypomethylated probes
}
