% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tr_tad_tables.R
\name{top_tr_tad_tables}
\alias{top_tr_tad_tables}
\title{top_tr_tad_tables}
\usage{
top_tr_tad_tables(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  top_gene_number,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step6 results from the top_tr_tabulation function. This function will also create a new step7 folder there if it has not been created, with a subdirectory with 'scatterplots' containing subfolders the results for the top genes and top TFs separately.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create TAD tables for the probes linked to the top genes/TRs by most hypermeth probes with G+ links.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create TAD tables for the probes linked to the top genes/TRs by most hypermeth probes with G- links.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create TAD tables for the probes linked to the top genes/TRs by most hypometh probes with G+ links.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create TAD tables for the probes linked to the top genes/TRs by most hypometh probes with G- links.}

\item{top_gene_number}{Specify a number to generate TAD tables for the probes linked to that many of the top genes/TFs based on the most linked enhancer probes.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Exports .tsv files with rows for each of the probes linked to the top X genes/TFs as specified by the user, with the probe locations in the hg38 human genome, which of the top X genes/TFs they're linked to, and the number of genes as well as their ensembl IDs and names of all the genes found in the same TAD as the probe for each of the user input TAD files.
}
\description{
This is a step7 function of the TENETR package.
This function takes the top genes/TFs by number of linked probes identified from
the step6 top_tr_tabulation function up to the number as specified by the user
and generates tables of information noting which genes/TFs each probe is linked to,
as well as the total number of genes, as well as their names, located within the same
TAD of each probe in each of the user-supplied TAD files.
}
