% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tr_ucsc_bed_files.R
\name{top_tr_ucsc_bed_files}
\alias{top_tr_ucsc_bed_files}
\title{top_tr_ucsc_bed_files}
\usage{
top_tr_ucsc_bed_files(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  top_gene_number,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step6 results from the top_tr_tabulation function. This function will also create a new step7 folder there if it has not been created, with a subdirectory with 'scatterplots' containing subfolders the results for the top genes and top TFs separately.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create bed files that can be uploaded to the UCSC genome browser showing links between the top genes/TRs by most hypermeth probes with G+ links, and said linked probes.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create bed files that can be uploaded to the UCSC genome browser showing links between the top genes/TRs by most hypermeth probes with G- links, and said linked probes.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create bed files that can be uploaded to the UCSC genome browser showing links between the top genes/TRs by most hypometh probes with G+ links, and said linked probes.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create bed files that can be uploaded to the UCSC genome browser showing links between the top genes/TRs by most hypometh probes with G- links, and said linked probes.}

\item{top_gene_number}{Specify a number to generate TAD tables for the probes linked to that many of the top genes/TFs based on the most linked enhancer probes.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Exports .bed formatted files that act as an interact file for uploading to the UCSC genome browser see: https://genome.ucsc.edu/goldenPath/help/interact.html. These files display the interactions between the top genes/TFs as noted by the user, and each of the linked enhancer probes to each individual gene for given analysis types.
}
\description{
This is a step7 function of the TENETR package.
This function takes the top genes/TFs by number of linked probes identified from
the step6 top_tr_tabulation function up to the number as specified by the user
and generates bed files that can be uploaded to the UCSC genome browser displaying the
links between each of the top specified genes/enhancers of the given analysis types
and their linked enhancer probes.
}
