% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tr_simple_scatterplots.R,
%   R/top_tr_survival.R
\docType{data}
\name{TENET_directory}
\alias{TENET_directory}
\title{top_tr_simple_scatterplots}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
TENET_directory

TENET_directory
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step6 results from the top_tr_tabulation function. This function will also create a new step7 folder there if it has not been created, with a subdirectory called histogram containing the results.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create histograms for the top TRs by most hypermeth probes with G+ links.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypermeth probes with G- links.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypometh probes with G+ links.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create histograms for the top TRs by most hypometh probes with G- links.}

\item{top_gene_number}{Specify a number to generate survival plots for that many of the top genes and TFs based on the most linked enhancer probes.}

\item{visualize_survival_plots_genes}{Set to TRUE/FALSE depending on if you want to create .pdfs displaying the survival results for the genes of interest.}

\item{visualize_survival_plots_probes}{Set to TRUE/FALSE depending on if you want to create .pdfs displaying the survival results for the probes linked to the genes of interest.}

\item{high_thresh}{Set a number ranging from 0 to 1, as a threshold for proportion of samples to include in the high expression/methylation group, and should be greater than low_thresh to prevent samples from appearing in both groups.}

\item{low_thresh}{Set a number ranging from 0 to 1, as a threshold for proportion of samples to include in the low expression/methylation group, and should be less than low_thresh to prevent samples from appearing in both groups.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns .pdf files with the scatterplots showing the expression of the gene of interest on the x-axis and the methylation of the linked probe on the y-axis.

Returns dataframes of survival information in the form of .tsv files, as well as .pdfs if selected by the user, showing survival information for the expression of the top genes, as well as the probes linked to them.
}
\description{
This is a step7 function of the TENETR package.
This function takes the top TRs by number of linked probes identified from
the step6 top_tr_tabulation function up to the number as specified by the user
and generates scatterplots displaying the expression level of each of these genes
and the methylation level of each enhancer probe linked to them.

This is a step7 function of the TENETR package.
This function takes the top TRs by number of linked probes identified from
the step6 top_tr_tabulation function up to the number as specified by the user
and generates survival plots for the expression level of each gene as well
as the DNA methylation of each enhancer probe linked to them.
}
\keyword{datasets}
