% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tr_circos.R
\name{top_tr_circos}
\alias{top_tr_circos}
\title{top_tr_ciros}
\usage{
top_tr_circos(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis,
  top_gene_number,
  core_count
)
}
\arguments{
\item{TENET_directory}{Set a path to the directory that contains step6 results from the top_tr_tabulation function. This function will also create a new step7 folder there if it has not been created, with a subdirectory with 'circos' containing the results.}

\item{hypermeth_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to create circos plots for the top genes/TRs by most hypermeth probes with G+ links to those probes.}

\item{hypermeth_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create circos plots for the top genes/TRs by most hypermeth probes with G- links to those probes.}

\item{hypometh_Gplus_analysis}{Set to TRUE/FALSE depending on if you want to to create circos plots for the top genes/TRs by most hypometh probes with G+ links to those probes.}

\item{hypometh_Gminus_analysis}{Set to TRUE/FALSE depending on if you want to to create circos plots for the top genes/TRs by most hypometh probes with G- links to those probes.}

\item{top_gene_number}{Specify a number to generate circos plots for that many of the top genes/TFs based on the most linked enhancer probes.}

\item{core_count}{Argument passed as mc.cores argument for mclapply. See ?mclapply from the parallel package for more details.}
}
\value{
Currently returns .html files with the circos plots displaying the top genes and TRs and links to their linked enhancer probes.
}
\description{
This is a step7 function of the TENETR package.
This function takes the top genes/TRs by number of linked probes identified from
the step6 top_tr_tabulation function up to the number as specified by the user
and generates circos plots for each gene showing the genomic links between
each gene and each probe linked to it.
}
